chore: refactor identity API, adding POST /quartz/orgs and GET /clusters
